+++ src/exe_backend.Contract/Services/Subscription/Command.cs
@@ -2,5 +2,5 @@ namespace exe_backend.Contract.Services.Subscription;
 
 public static class Command
 {
-    public record CreateSubscriptionPackageCommand(string Name, long Price, int ExpiredMonth, string? Description = null) : ICommand;
+    public record CreateSubscriptionPackageCommand(string Name, int Price, int ExpiredMonth, string? Description = null) : ICommand;
 }
